requirements.txt
ccxt==4.1.13
pandas==2.1.4
pandas-ta==0.3.14b0
python-telegram-bot==20.7
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
ta-lib==0.4.28
tensorflow==2.15.0
yfinance==0.2.36
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0 